+++ R/FamiliarEnsemble.R
@@ -138,38 +138,6 @@ setMethod("show", signature(object="familiarEnsemble"),
 
 
 
-#####assess_calibration (ensemble)#####
-setMethod("assess_calibration", signature(object="familiarEnsemble"),
-          function(object, data, eval_times=NULL, is_pre_processed=FALSE){
-            
-            # Load eval_times from the object settings attribute, if it is not provided.
-            if(is.null(eval_times)){
-              eval_times <- object@settings$eval_times
-            }
-            
-            # Check eval_times argument
-            if(object@outcome_type %in% c("survival")){
-              sapply(eval_times, .check_number_in_valid_range, var_name="eval_times", range=c(0.0, Inf), closed=c(FALSE, TRUE))
-            }
-            
-            # Test if models are properly loaded
-            if(!is_model_loaded(object=object)){
-              ..error_ensemble_models_not_loaded()
-            }
-            
-            # This function is the same for familiarModel and familiarEnsemble objects
-            return(.assess_calibration(object=object, data=data, eval_times=eval_times, is_pre_processed=is_pre_processed))
-          })
-
-#####compute_calibration_data (ensemble, dataObject)#####
-setMethod("compute_calibration_data", signature(object="familiarEnsemble", data="dataObject"),
-          function(object, data, time=NULL){
-            
-            # This function is the same for familiarModel and familiarEnsemble objects
-            return(.compute_calibration_data(object=object, data=data, time=time))
-          })
-
-
 .get_available_stratification_ensemble_methods <- function(){
   return(c("ensemble_mode", "ensemble_mean", "mean_threshold", "median_threshold"))
 }